Add conservative flux smoothing option to divflux (forward thk + data_assimilation)
New smooth_sigma argument in compute_divflux and
compute_divflux_slope_limiter: when > 0, the staggered fluxes Qx/Qy
are Gaussian-filtered (std in cells, kernel spanning +/- 3 sigma)
BEFORE the divergence is taken. The result is the exact discrete
divergence of a (filtered) flux, so the mass budget is preserved --
unlike smoothing divflux after the fact. This tames the grid-scale
noise that differentiation amplifies (emulator jitter, upwind
switching, rough data-based geometry) while passing the
long-wavelength signal untouched.

Exposed as two config keys (both default 0.0 = off, previous
behaviour bitwise unchanged):

- processes.thk.divflux_smooth_sigma: filters the transport fluxes in
  the stock thk update; the filtered flux IS the transport operator,
  so conservation holds by construction. Validated on Aletsch (100 m):
  at sigma=1.5 the grid-scale divflux rms of the initial
  data-based-geometry shock drops from 44 to 14 m/yr (emulator floor
  ~0.5 m/yr after 10 yr) with volume drift < 0.15%; sigma > ~3 starts
  to remove resolved flux structure.

- assimilations.data_assimilation.divflux.smooth_sigma: same filter in
  the divflux diagnostic and in all divflux-based cost terms
  (cost_divfluxobs, cost_divfluxfcz, cost_divfluxpen), for both
  optimizers, so the inversion residuals see the same operator.

Also: gaussian_filter_tf now casts its kernel to the input dtype
(float64-safe).

+++ igm/utils/math/gaussian_filter_tf.py
@@ -12,6 +12,6 @@ def gaussian_kernel(size: int, sigma: float):
 def gaussian_filter_tf(qx, sigma=1.0, kernel_size=5):
     """Apply 2D Gaussian filter to qx using TensorFlow."""
     qx = qx[tf.newaxis, :, :, tf.newaxis]  # shape (1, ny, nx, 1)
-    kernel = gaussian_kernel(kernel_size, sigma)
+    kernel = tf.cast(gaussian_kernel(kernel_size, sigma), qx.dtype)
     qx_smooth = tf.nn.conv2d(qx, kernel, strides=1, padding='SAME')
     return qx_smooth[0, :, :, 0]  # shape (ny, nx)
